+++ sdk/storage/azure_storage_blob/src/clients/page_blob_client.rs
@@ -18,25 +18,25 @@ impl PageBlobClient {
     ///
     /// # Arguments
     ///
-    /// * `endpoint` - The full URL of the Azure storage account, for example `https://myaccount.blob.core.windows.net/`
+    /// * `account_url` - The full URL of the Azure storage account, for example `https://myaccount.blob.core.windows.net/`
     /// * `container_name` - The name of the container containing this Page blob.
     /// * `blob_name` - The name of the Page blob to interact with.
     /// * `credential` - An optional implementation of [`TokenCredential`] that can provide an Entra ID token to use when authenticating.
     /// * `options` - Optional configuration for the client.
     pub fn new(
-        endpoint: &str,
+        account_url: &str,
         container_name: &str,
         blob_name: &str,
         credential: Option<Arc<dyn TokenCredential>>,
         options: Option<PageBlobClientOptions>,
     ) -> Result<Self> {
-        let mut url = Url::parse(endpoint)?;
+        let mut url = Url::parse(account_url)?;
 
         {
             let mut path_segments = url.path_segments_mut().map_err(|_| {
                 azure_core::Error::with_message(
                     azure_core::error::ErrorKind::Other,
-                    "Invalid endpoint URL: Failed to parse out path segments from provided endpoint URL.",
+                    "Invalid account URL: Failed to parse out path segments from provided account URL.",
                 )
             })?;
             path_segments.extend([container_name, blob_name]);