+++ sdk/storage/azure_storage_queue/examples/queue_service_client.rs
@@ -42,11 +42,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let account = env::var("AZURE_QUEUE_STORAGE_ACCOUNT_NAME")
         .expect("Set AZURE_QUEUE_STORAGE_ACCOUNT_NAME environment variable");
 
-    let endpoint = format!("https://{}.queue.core.windows.net/", account);
+    let account_url = format!("https://{}.queue.core.windows.net/", account);
     let queue_name = random_queue_name();
 
     let credential = DeveloperToolsCredential::new(None)?;
-    let service_client = QueueServiceClient::new(&endpoint, Some(credential.clone()), None)?;
+    let service_client = QueueServiceClient::new(&account_url, Some(credential.clone()), None)?;
     let queue_client = service_client.queue_client(&queue_name)?;
 
     println!("Creating queue '{queue_name}'...");